Pass level 6

+++ src/StringUtils.php
@@ -59,7 +59,7 @@ class StringUtils
 
         $targetPosition = null;
         $targetPositionOffset = 0;
-        $targetLength = mb_strlen($target, $encoding);
+        $targetLength = (int) mb_strlen($target, $encoding);
 
         $mutableFragment = $fragment;
